Log level and Export Logs features in new details view
This feature enables a user to set the Ditto logging level, or to disable logging.  The setting
is stored in UserDefaults.standard, and when Ditto is restarted, logging begins from the last
stored level; the default is .disabled. This featue introduces a LoggingDetailsView which hosts
the new log level setting and the existing export logs feature.

- Ditto log level user setting and export logs in new dedicated UI