Heartbeat feature 1st full draft
This commit adds a simple UI for displaying the output of the hearbeat tool.
In developing the view a number of issues with the tool came to light, among
which was that the output of the HeartbeatInfo class did not match data written
to the collection, leading to a substantial refactor of the tool.

- refactor HeartbeatVM to normalize data output from the callback and to
the (devices) collection
- add new output-related data classes/structs
- add HeartbeatView and related supporting structures and extensions to
visualize a list of heartbeat documents from the configured collection