Update peers list (#76)
* Remove outdated code

* use peerKey to identify device

Use peerKey instead of addressSiteId. The peerKey is a unique identifier to each device (peer). The peerKey is printed in logs which will make it easier to identify a device in the logs.

* use peerKey to identify device

Use peerKey instead of addressSiteId. The peerKey is a unique identifier to each device (peer). The peerKey is printed in logs which will make it easier to identify a device in the logs.

* format code with extensions files

* Remove unused imports

* Fix wrong peerKey in PeersListView

- fix wrong peerKey in PeersListView
- add static and instance peerKey accessor to DittoPeer extension
- remove Data extension

---------

Co-authored-by: Eric Turner <eric.turner@ditto.live>